Flyweight
(Epilepsy warning if you try to run the code - the flashing colors can be commented out in UpdateSharedModel)

An implementation of the Flyweight model. A number of gameobjects are created from one shared data class.

Because of how Unity's component based Gameobjects, each instance still represents its own GameObject and SpriteRenderer which have to be handled seperately, making the memory saving purpose of Flyweight somewhat ineffective. In my example, I focused on making identical, connected gameobjects for an effect. The result is basically a Unity prefab created only in code, with the benefit that each GameObject instance has no code/script attatched to it.